+++ server/src/types/error.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
-import { HttpCode } from '../types/httpCode';
+import { HttpCode } from '../types';
 
 /**
  * ApiError class represents a custom API related error, extends the native Error class. Includes
