+++ server/src/types/index.ts
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
+//
+// SPDX-License-Identifier: MIT
+
+export * from './error';
+export * from './formulas';
+export * from './general';
+export * from './grades';
+export * from './httpCode';
+export * from './language';
+export * from './model';
+export * from './sisu';
