+++ server/test/controllers/assessmentModel.test.ts
@@ -8,7 +8,7 @@ import TeacherInCharge from '../../src/database/models/teacherInCharge';
 
 import { mockTeacher } from '../mock-data/misc';
 import { app } from '../../src/app';
-import { HttpCode } from '../../src/types/httpCode';
+import { HttpCode } from '../../src/types';
 import { Cookies, getCookies } from '../util/getCookies';
 
 const request: supertest.SuperTest<supertest.Test> = supertest(app);
