+++ server/test/controllers/auth.test.ts
@@ -8,9 +8,9 @@ import supertest, { SuperAgentTest } from 'supertest';
 
 import { JWT_COOKIE_EXPIRY_MS, JWT_EXPIRY_SECONDS } from '../../src/configs/constants';
 
-import { SystemRole } from 'aalto-grades-common/types/auth';
+import { SystemRole } from 'aalto-grades-common/types';
 import { app } from '../../src/app';
-import { HttpCode } from '../../src/types/httpCode';
+import { HttpCode } from '../../src/types';
 
 const request: supertest.SuperTest<supertest.Test> = supertest(app);
