+++ server/test/controllers/course.test.ts
@@ -5,7 +5,7 @@
 import supertest from 'supertest';
 
 import { app } from '../../src/app';
-import { HttpCode } from '../../src/types/httpCode';
+import { HttpCode } from '../../src/types';
 import { Cookies, getCookies } from '../util/getCookies';
 
 const request: supertest.SuperTest<supertest.Test> = supertest(app);
