+++ server/test/functions/csvParsing.test.ts
@@ -4,9 +4,7 @@
 
 import { AttainmentGradeData } from 'aalto-grades-common/types';
 import { parseGradesFromCsv, parseHeaderFromCsv } from '../../src/controllers/grades';
-import { ApiError } from '../../src/types/error';
-import { StudentGrades } from '../../src/types/grades';
-import { HttpCode } from '../../src/types/httpCode';
+import { ApiError, HttpCode, StudentGrades } from '../../src/types';
 
 function checkError(error: unknown, httpCode: HttpCode, message: string | Array<string>): void {
   expect(error).toBeInstanceOf(ApiError);
