+++ server/test/util/getCookies.ts
@@ -5,7 +5,7 @@
 import supertest from 'supertest';
 
 import { app } from '../../src/app';
-import { SystemRole } from 'aalto-grades-common/types/auth';
+import { SystemRole } from 'aalto-grades-common/types';
 
 const request: supertest.SuperTest<supertest.Test> = supertest(app);
